+++ ChangeLog.md
@@ -5,6 +5,7 @@
    2. Health Potion
    3. Repair Kit
    4. Smoke Bomb
+   5. Explosion Potion
 - Reworked *Urgent* weapon variant: it now adds priority to the move
 - Renamed *Swift* weapon variant to *Accelerating*